Move player files from client to model
